src/xfmr_zem/servers/doc_parser/requirements.txt
pymupdf
landingai-ade
pillow
Ipython
dotenv
PyYAML
matplotlib
seaborn
numpy<2.0.0
transformers>=4.55.2
torchvision
torch>=2.5.0
rich
#uv pip install paddlepaddle-gpu==3.2.1 -i https://www.paddlepaddle.org.cn/packages/stable/cu126/
#uv pip install -U paddleocr[doc-parser]